price data finalized, cleaning tomorrow
