+++ data/ingredient_prices_clean.csv
@@ -1,0 +1,3 @@
+,hot sauce,soda,prosecco,fruit juice,butter,advocaat,flavored rum,erin cream,tequila,milk,pernod,water,liqueur,drambuie,aquavit,cornstarch,sugard,nut,galliano,flavored vodka,cognac,ice,bourbon,vermouth,ice cream,sambuca,fruit,zima,baileys,brandy,coffee,grain alcohol,mix,spice,tonic water,vodka,port,herb,beer,bitters,carbonated water,champagne,cherry heering,club soda,egg,food coloring,red wine,soy sauce,sweet and sour,wine,pisang ambon,dry vermouth,glycerine,everclear,rum,olive brine,whiskey,cachaca,gin,schnapps,sarsaparilla,jagermeister,absinthe,sherry,scotch,sugar,frangelico,pisco,ricard,aperitif,dubonnet rouge
+Unnamed: 0,hot sauce,soda,prosecco,fruit juice,butter,advocaat,flavored rum,erin cream,tequila,milk,pernod,water,liqueur,drambuie,aquavit,cornstarch,sugard,nut,galliano,flavored vodka,cognac,ice,bourbon,vermouth,ice cream,sambuca,fruit,zima,baileys,brandy,coffee,grain alcohol,mix,spice,tonic water,vodka,port,herb,beer,bitters,carbonated water,champagne,cherry heering,club soda,egg,food coloring,red wine,soy sauce,sweet and sour,wine,pisang ambon,dry vermouth,glycerine,everclear,rum,olive brine,whiskey,cachaca,gin,schnapps,sarsaparilla,jagermeister,absinthe,sherry,scotch,sugar,frangelico,pisco,ricard,aperitif,dubonnet rouge
+0,0.626,0.0346874999999999,0.5315352615287554,0.1027118644067796,0.776875,,1.485,5.894375,0.6596873090041601,0.0617578125,,0.9966666666666668,,,,0.101875,,0.4985714285714286,99.5,0.3966798761854065,2.2633962264150944,0.4683333333333333,,,0.2688541666666666,,1.37,,,,,0.2466666666666666,0.1105250653361843,0.8787061994609164,0.6860416666666667,0.391844835566371,,5.7325,1.7066666666666666,1.7475,0.0820610584717486,0.3263769773203736,0.2446173570019723,0.0203550295857988,0.7956009615384615,1.4264593931280676,0.3032931790251243,0.15645,0.8070442719587156,0.2879477928274285,,,,,,,,,,,,,,,,,,,,,
