+++ data/ingredient_prices_raw.csv
@@ -1,0 +1,72 @@
+,0
+hot sauce,0.626
+soda,0.034687499999999996
+prosecco,0.5315352615287554
+fruit juice,0.10271186440677965
+butter,0.776875
+advocaat,
+flavored rum,1.485
+erin cream,5.894375
+tequila,0.6596873090041601
+milk,0.0617578125
+pernod,
+water,0.9966666666666667
+liqueur,
+drambuie,
+aquavit,
+cornstarch,0.101875
+sugard,
+nut,0.4985714285714286
+galliano,99.5
+flavored vodka,0.3966798761854065
+cognac,2.2633962264150944
+ice,0.4683333333333333
+bourbon,
+vermouth,
+ice cream,0.26885416666666667
+sambuca,
+fruit,1.37
+zima,
+baileys,
+brandy,
+coffee,
+grain alcohol,0.24666666666666667
+mix,0.11052506533618439
+spice,0.8787061994609164
+tonic water,0.6860416666666667
+vodka,0.39184483556637106
+port,
+herb,5.7325
+beer,1.7066666666666666
+bitters,1.7475
+carbonated water,0.08206105847174862
+champagne,0.3263769773203736
+cherry heering,0.24461735700197237
+club soda,0.020355029585798816
+egg,0.7956009615384615
+food coloring,1.4264593931280676
+red wine,0.30329317902512437
+soy sauce,0.15645
+sweet and sour,0.8070442719587156
+wine,0.2879477928274285
+pisang ambon,
+dry vermouth,
+glycerine,
+everclear,
+rum,
+olive brine,
+whiskey,
+cachaca,
+gin,
+schnapps,
+sarsaparilla,
+jagermeister,
+absinthe,
+sherry,
+scotch,
+sugar,
+frangelico,
+pisco,
+ricard,
+aperitif,
+dubonnet rouge,
